// plio-api/plio/apis/AssetController.ts
/* eslint-disable */

/**
 * Plio.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { createEndpoint, normalizePath } from '@1023-ventures/darri-core';

import {
    Asset,
    AssetFromJSON,
    AssetToJSON,
} from '../models/Asset';
import {
    ImageData,
    ImageDataFromJSON,
    ImageDataToJSON,
} from '../models/ImageData';
import {
    ProblemDetails,
    ProblemDetailsFromJSON,
    ProblemDetailsToJSON,
} from '../models/ProblemDetails';

// Data Transfer Objects for individual endpoints

type DeleteUrlParams = {
    /** id  true*/
    id: string;
}

type DeleteQueryParams = {
    /**  */
    concurrency?: string;
}

type CreateBody =
    /**  */
    & Asset;

type CreateUrlParams = {
    /** reload  true*/
    reload: boolean;
}

type ReadUrlParams = {
    /** id  true*/
    id: string;
}

type SparseUpdateBody =
    /**  */
    & string;

type SparseUpdateUrlParams = {
    /** id  true*/
    id: string;
    /** reload  true*/
    reload: boolean;
}

type UpdateBody =
    /**  */
    & Asset;

type UpdateUrlParams = {
    /** reload  true*/
    reload: boolean;
}

type UploadImageBody =
    /**  */
    & ImageData;

/**
 * no description
 */
export const AssetController = {

    /**
     */
    _delete: createEndpoint<boolean, never, DeleteUrlParams, DeleteQueryParams>({
        method: 'DELETE',
        url: normalizePath('/plio/api/Asset/{id}')
    }),
    /**
     */
    create: createEndpoint<Asset, CreateBody, CreateUrlParams, never>({
        method: 'POST',
        url: normalizePath('/plio/api/Asset')
    }),
    /**
     */
    read: createEndpoint<Asset, never, ReadUrlParams, never>({
        method: 'GET',
        url: normalizePath('/plio/api/Asset/{id}')
    }),
    /**
     */
    readAll: createEndpoint<Array<Asset>, never, never, never>({
        method: 'GET',
        url: normalizePath('/plio/api/Asset')
    }),
    /**
     */
    sparseUpdate: createEndpoint<Asset, SparseUpdateBody, SparseUpdateUrlParams, never>({
        method: 'PATCH',
        url: normalizePath('/plio/api/Asset/{id}')
    }),
    /**
     */
    update: createEndpoint<Asset, UpdateBody, UpdateUrlParams, never>({
        method: 'PUT',
        url: normalizePath('/plio/api/Asset')
    }),
    /**
     */
    uploadImage: createEndpoint<unknown, UploadImageBody, never, never>({
        method: 'POST',
        url: normalizePath('/plio/api/Asset/uploadImage')
    }),
}